+++ Data/SBV/Provers/Prover.hs
@@ -92,7 +92,7 @@ mkConfig s smtVersion startOpts = SMTConfig { verbose                     = Fals
                                             , dsatPrecision               = Nothing
                                             , extraArgs                   = []
                                             , satCmd                      = "(check-sat)"
-                                            , satTrackUFs                 = True                   -- i.e., yes, do extract UI function values
+                                            , allSatTrackUFs              = True                   -- i.e., yes, do extract UI function values
                                             , allSatMaxModelCount         = Nothing                -- i.e., return all satisfying models
                                             , allSatPrintAlong            = False                  -- i.e., do not print models as they are found
                                             , isNonModelVar               = const False            -- i.e., everything is a model-variable by default