+++ Documentation/SBV/Examples/Puzzles/Garden.hs
@@ -90,4 +90,4 @@ puzzle = do n <- sInteger "N"
 -- over the prefix existentials and the assignments to uninterpreted function 'col'
 -- for model purposes here, as they don't represent a different solution.)
 flowerCount :: IO ()
-flowerCount = print =<< allSatWith z3{satTrackUFs=False} puzzle
+flowerCount = print =<< allSatWith z3{allSatTrackUFs=False} puzzle