+++ SBVBenchSuite/BenchSuite/Puzzles/Garden.hs
@@ -26,4 +26,4 @@ import BenchSuite.Bench.Bench as S
 -- benchmark suite
 benchmarks :: Runner
 benchmarks = rGroup [ S.runWith s "Garden" puzzle `using` runner allSatWith ]
-  where s = z3{satTrackUFs = False, isNonModelVar = ("_modelIgnore" `isSuffixOf`)}
+  where s = z3{allSatTrackUFs = False, isNonModelVar = ("_modelIgnore" `isSuffixOf`)}